pyproject.toml
[tool.poetry]
name = "lightbulb.ext.hecate"
version = "0.1.5"
description = "File driven wrappers around lightbulb extensions and plugins."
authors = ["Radkii <real.radkii@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/radkii/lightbulb-ext-hecate"
packages = [{include = "lightbulb\\ext\\hecate"}]

[tool.poetry.dependencies]
python = ">=3.9 <3.12"
hikari-lightbulb = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
